pyproject.toml
[project]
name = "sre-incident-env"
version = "1.0.0"
description = "SRE Incident Response OpenEnv — a simulated broken Linux server for AI agent training"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.135.3",
    "uvicorn==0.44.0",
    "pydantic==2.12.5",
    "requests==2.33.1",
    "openai==2.30.0",
    "httpx==0.28.1",
    "python-multipart==0.0.24",
    "openenv-core==0.2.3",
    "fastmcp==3.2.0",
]

[project.scripts]
serve = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
